Speed up StackTraceFrame::GetFileName()
This retrieves script name directly from StackFrameBase, bypassing
building of StackFrameInfo if one hasn't already been initialized,
thus avoiding computation of expensive properties that are not
required. This matches current behavior of GetScriptNameOrSourceURL()
and is a workaround until a dedicated API is available.

This is necessary to switch AdTagging over from using
GetScriptNameOrSourceURL() to GetScriptName(), to ensure that
scripts with source urls are tagged appropriately.

Change-Id: I6eb145b88c26deb1a088f038b0f8b377bc8fe3ab