[TurboFan] Experiment to remember field offsets for deopts
It would be great if we can avoid deoptimizing if a field lost its
constness for a field that isn't referenced by a particular optimized
function. This code remembers the field indexes for the first 32
fields.

In a performance run on Octane, with 39 deoptimizations, it only reduced
deopts by one, so it's not necessarily useful in that scenario.

Bug: v8:11163
Change-Id: I9c7f07c64f5deafd902cf27049ca09c6cdd3d179