[test][turbofan] Add %ObserveNode intrinsic for node-specific tests
(Copied from nicohartmann@ CL
https://chromium-review.googlesource.com/c/v8/v8/+/2135631)

This CL adds a new intrinsic %ObserveNode(expr) which has noop semantics
but triggers the new NodeObserver set on the OptimizedCompilationInfo
when the node generated for expr is created or changed in any phase
(until SL).

This provides the infrastructure to write reasonable unit tests that
check for the construction of or lowering to specific nodes (e.g.
depending on feedback).

A first use case (cctest/test-sloppy-equality.cc) is included in this CL.

Some points are still todo:
 * observation of other node modifications (replacement, change of
   inputs, ...)
 * move end of observation further down the pipeline (e.g. ECL)

Change-Id: Idc5a5e38af8b1d9a2ec5021bf821c4e4e1406220