[wasm-simd][scalar-lowering] Fix lowering of load nodes
We forgot to check if a load's input node (index) has any replacement.
This led to weird cases like I32x4ExtractLane persisting even after
scalar lowering is done, which is incorrect.

This manifested in a crash, where we try to call pextrd with a general
register operand.

With this, we can run all currently checked in performance tests without
crashing.

Bug: chromium:1124885
Change-Id: Ide36ef94ab5f63446c725b9c2eb64be01e7fa6ab