[wasm][inspector][test] Print number of script IDs
We currently do not report a script ID for wasm code, i.e. the script id
is 0. It's thus the same as e.g. for the wasm-to-js wrapper.
We cannot print the script IDs themselves, as they are considered
unstable. But by printing the number of script IDs we can check (later)
that a different script id is being reported for wasm code.

R=thibaudm@chromium.org

Bug: chromium:1125986
Change-Id: Ibc52829ea8a5a5c9506e36390eb4c608bcab4624