[api] Store initialized lazy accessor directly in AccessorPair
For API objects, we lazily initialize JSFunctions for accessors (only
when Object.getOwnPropertyDescriptor is called) to save memory.

To prevent creating different JSFunctions for the same accessor, we
store the initialized JSFunctions in the template cache.

This generally works ok but the template cache is actually for fast
creation of commonly instantiated objects, not for these
accessors. The template cache is split into two levels for even faster
access and these getters usually take up the first level of the cache
due to the order in which they are created by Blink.

Rather than (mis)using the template cache for these JSFunctions, we
replace the original FunctionTemplateInfo directly in the AccessorPair
with the newly instantiated JSFunction as there is no need for the
FunctionTemplateInfo anymore.

The runtime can handle either a FunctionTemplateInfo or a JSFunction
in an AccessorPair perfectly fine already so there isn't any other
change required.

A future CL will disable the caching of these JSFunctions in the
template cache.

Bug: v8:11284
Change-Id: Ib2b2987e2a5908d03b55ac2dc99fc83184e83c09