[wasm-simd][liftoff][x64] Move v128.select into macro-assembler
This allows us to reuse this optimized code sequence in Liftoff.

We can't do the same thing in IA32 yet, there is no kScratchDoubleReg
defined in the macro-assembler-ia32.cc, it is defined in code-generator-ia32
as xmm0 but I'm not sure if it is safe to just use that in the macro assembler.

Change-Id: I6c761857c49d2518fbc82cd0796c62fc86665cb5