[tools] Fix ic-processor-driver.mjs
The IC object's interface is changing all the time and this code is
just bitrotting. Rather than trying to keep this updated all the time,
let's just use Object.values to print all the key value pairs in the
ic object.

This looks slightly worse than the previous text format but it has the
critical advantage of being broken less often.

Change-Id: Ia3580d1ba82a981d8442682f66d6002436e70f42