[tools] Add DisableGCMole scope
Make sure gcmole detects issue in DisallowHeapAllocation scopes.

DisallowHeapAllocation is widely used in the codebase to document code
that doesn't allocate. However, this has the rather unexpected
side-effect that gcmole is not run when such a scope is active.

This CL changes the default behavior of gcmole to run even with
DisallowHeapAllocation scopes present. This will give us the best
results of both worlds, dynamic checks by the fuzzer, and static
analysis by gcmole.

To allow crazy local raw pointer operations there is a new
DisableGCMole scope that explicitely disables gcmole.

Change-Id: I0a78fb3b4ceaad35be9bcf7293d917a41f90c91f