Merged: [wasm-simd][ia32] Fix instruction operand constraints for ext mul
Ext mul's codegen assumes that all inputs are in registers, but the
instruction-selector wasn't the correct constraints. The codegen for ext
mul is slightly complicated so we chose to restrict the inputs to be
registers rather than changing codegen.

Bug: chromium:1165966,v8:11262
TBR=bbudge@chromium.org
(cherry picked from commit 23816fcb5851b501e52aedc7929fb953052a248e)

Change-Id: I916cad6efe7f5535bae56695d9847bac588975b2