[TurboFan] Move FeedbackVectorRef to the never serialized list
Actual FeedbackVector IC slots don't participate in the Ref class, since
they are read during serialization as we process bytecode. So
FeedbackVectorRef really only deals with the FeedbackCell array and
the SharedFunctionInfo. These two types are already in the no serialized
list, so it's pretty easy to move this class over there too.

Bug:v8:7790
Change-Id: I51b7bf4c3404ae5bcfb16d29b5e719787ddd6b17