S390: Pass the value of SupportsWasmSimd128() to builtin
WasmCompileLazy needs to save the content of vector
parameter registers. If Simd is not enabled or the hardware
does not support Simd operations then we need to saves the value of
Double registers instead, therefore we need a way to pass the
value of "CpuFeatures::SupportsWasmSimd128()" to the builtin
during runtime.

Change-Id: I38883d6e3127b90d43c12499745936c2902be417