[turboprop] Don't consider stores to constant fields as mutable
Turboprop doesn't use optimizations based on field constness to reduce
the number of deoptimizations. While this is safe for loads, for stores
if a different value is stored to a const field we should update the
constness of the field. This is needed so we can safely deopt any other
code that is relying on the constness of the field. Currently, turboprop
doesn't do this. So for now treat stores to constant fields similar to
TurboFan. In future, we may consider adding code to update the field
constness if necessary to reduce the number of deoptimizations.

This also introduces a runtime-test function to check if
turboprop-as-toptier is set and updates %OptimizeFunctionOnNextCall
to update tier ups till the function reaches top tier when
%OptimizeFunctionOnNextCall is called on mid-tier optimized functions.


Bug: chromium:1172797
Change-Id: I1d660457cb5d647e1283a495040a7e452fe1ac7e